single trial analyses

+++ scripts/step06_SPMsingletrial/s02_rename_beta.py
@@ -13,7 +13,8 @@ main_dir = Path(current_dir).parents[1] # discovery: /dartfs-hpc/rc/lab/C/CANlab
 meta_dir = os.path.join(main_dir, 'data', 'dartmouth', 'd06_singletrial_SPM')
 spm_dir = os.path.join(main_dir, 'analysis', 'fmri', 'spm', 'multivariate', 's01_singletrial')
 nifti_dir = os.path.join(main_dir, 'analysis', 'fmri', 'spm', 'multivariate', 's02_isolatenifti')
-sub_list = next(os.walk(spm_dir))[1]
+sub_folders = next(os.walk(spm_dir))[1]
+sub_list = [i for i in sub_folders if i.startswith('sub-')]
 items_to_remove = ['sub-0001','sub-0002','sub-0003','sub-0004','sub-0005',
 'sub-0006','sub-0007','sub-0008','sub-0009','sub-0010',
 'sub-0014','sub-0015','sub-0016','sub-0018','sub-0019','sub-0020',
@@ -21,6 +22,7 @@ items_to_remove = ['sub-0001','sub-0002','sub-0003','sub-0004','sub-0005',
 'sub-0032', 
 'sub-0050','sub-0030', 'sub-0035',  'sub-0043', 
  'sub-0080','sub-0031','sub-0033','sub-0037','sub-0026','sub-0028', 'sub-0029' ]
+items_to_remove = ['singletrial_SPM_03-pain-post', 'singletrial_SPM_02-pain-late','singletrial_SPM_04-pain-plateau','singletrial_SPM_01-pain-early']
 for item in items_to_remove:
     if item in sub_list:
         sub_list.remove(item)
@@ -40,11 +42,14 @@ for sub in sub_list:
     for ind, row in subset.iterrows():
 #        print(ind, row)
         source_name = os.path.join(spm_dir, sub, row.source_name)
-        nifti_name = f"sub-{row['sub']:04d}_ses-{row['ses']:02d}_run-{row['run']:02d}-{row.task}_task-social_ev-{row['ev']}-{int(row['num']):04d}.nii"
-        dest_name = os.path.join(nifti_dir, sub, row.nifti_name + '.nii')
+        nifti_name = f"sub-{row['sub']:04d}_ses-{row['ses']:02d}_run-{row['run']:02d}-{row['task']}_task-social_ev-{row['ev']}-{int(row['num']):04d}.nii"
+        dest_name = os.path.join(nifti_dir, sub, nifti_name)
         Path(os.path.join(nifti_dir, sub)).mkdir(parents=True, exist_ok=True)
         print(source_name)
         print(dest_name)
-        shutil.copy(source_name, dest_name)
+        if os.path.exists(source_name):
+            shutil.copy(source_name, dest_name)
+        else:
+            break
 
 # %%
